requirements.txt
Django==4.2.27
PyJWT
django-environ
django-cors-headers
mysqlclient
PyMySQL
gunicorn
whitenoise
google-generativeai
requests